+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 * Fix some translations
 * Update st-page-title doc
 * Update radio menu component documentation
-
+* Create help component documentation
 
 
 ## 1.3.0 (Abril 4, 2017)